feat: update data types for book and library event IDs to Long and enhance Kafka consumer group configuration

+++ library-events-consumer-v2/src/main/java/com/learnkafka/controller/LibraryEventController.java
@@ -33,7 +33,7 @@ public class LibraryEventController {
 
     @GetMapping("/{libraryEventId}")
     public ResponseEntity<LibraryEventResponseDto> getLibraryEventById(
-            @PathVariable Integer libraryEventId) {
+            @PathVariable Long libraryEventId) {
         log.info("GET /v1/library-events/{}", libraryEventId);
         return libraryEventService.findById(libraryEventId)
                 .map(ResponseEntity::ok)
