+++ library-events-consumer-v2/src/main/java/com/learnkafka/dto/BookDto.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record BookDto(
         @NotNull
-        Integer bookId,
+        Long bookId,
 
         @NotBlank
         String bookName,
